backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File Upload & Storage
Pillow==10.2.0
cloudinary==1.38.0
django-storages==1.14.2
boto3==1.34.25

# Security
django-ratelimit==4.1.0
argon2-cffi==23.1.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
Faker==22.5.0

# Validation
pydantic==2.5.3
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.4.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.20.0
